+++ RadRefinements/Utilities/InstrumentType.cs
@@ -1,0 +1,19 @@
+﻿namespace RadRefinements
+{
+    internal enum InstrumentType
+    {
+        None,
+        Compass,
+        ChipLog,
+        Clock,
+        Quadrant,
+        SunCompass,
+        Inclinometer,
+        Barometer,
+        Thermometer,
+        Hygrometer,
+        WindCompass,
+        Anemometer,
+        Weathervane
+    }
+}