+++ RadRefinementsDLBridge/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("RadRefinements")]
-[assembly: AssemblyDescription("https://github.com/bryon82/SailwindRadRefinements")]
+[assembly: AssemblyTitle("RadRefinementsDLBridge")]
+[assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("raddude")]
-[assembly: AssemblyProduct("RadRefinements")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("RadRefinementsDLBridge")]
 [assembly: AssemblyCopyright("Copyright ©  2026")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("c0c6b307-93f0-4209-8594-7f911e48e709")]
+[assembly: Guid("fcba7640-5933-4bfa-bacc-41351db4788b")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("1.8.0.0")]
-[assembly: AssemblyFileVersion("1.8.0.0")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]